VettLLMServer/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-upstage
langchain-pinecone
python-dotenv
pymongo
pytz